Add files via upload
Phytoplankton blooms often follow nutrient enrichment.  Differences among lakes in light-absorbing dissolved organic carbon (DOC) may shift  bloom thresholds to higher nutrient loads and thereby increase resilience of lakes to enrichment. To explore this idea, we measured resilience to experimental enrichment of two lakes with contrasting DOC concentrations. We compared bloom thresholds in both lakes using a model of phytoplankton response to DOC and nutrients, a dynamic time series indicator of resilience, and two empirical measures of stochastic resilience, mean exit time and median survival time. For the dynamic indicator and ecosystem model the lake with higher DOC was more resilient to enrichment. However, the distributions overlapped for stochastic indicators of resilience of the two lakes. These analyses show that DOC interacts with mixing depth and zooplankton biomass to affect resilience. Strong contrasts in DOC and many observations are needed to discern effects of DOC on resilience to enrichment.